Finally fiured out how to make the ball travel at a reasonable speed throughout the game
I created a common data model to store how fast the ball should go, once it is has launched, and then I used that information to finally create the limiter code. The limiter works by adjusting the velocity based on its current X and Y values. The code was rather complex only because I had to account for negative values as they represented left and downward trajectories. I also wanted to support two different speeds and the range in between them. This way the game can start off slower and then ramp up to a faster, still reasonable to play with speed.