Based on wife's feedback I have slowed the ball down even more
This speed appears to enable a more enjoyable game. I think that some of this code can be refactored as a lot of it is very similar. Will consider tackling that in the future as part of a clean up effort. For the time being I do not want to get bogged down writing super clean code versus continuing deeper into the tutorial.