Sound effects have been added to the game.
The lesson introduced the concept of sounds effects. In this lesson they have you add a sound effect onto the ball, but I thought that it would be nice if the ball made a different sound effect when it hit the blocks or the paddle. To prevent some code duplication I added the concept of a common data model, this allows me to share the audio playing code between any class that is going to leverage it. Currently that is the Paddle and the Block.