No need to track block count once level first starts.
Originally if the level object was created prior to all of the blocks being created this count would be off. Which did not matter in reality because the number of blocks is unimportant until the first one is destroyed. The lesson spoke about having each block add to the counter upon initializing which would resolve that issue. But while thinking about a better way to execute that, versus incrementing a number N times, I realized that this update was completely useless.