Player Score is now a Singleton, also introducing a logic layer
The course wanted to introduce the concept of a singleton, so I decided that I would use my Common.Singleton code and create an instance using that. I then decided that I should expand this into introducing N-tier architecture into this game by adding a logic layer. In future updates I will move additional non-Unity specific logic over there.

One important thing to note here is how I implemented the PlayerScore. In the logic layer I created the PlayerScore class, but then I struggled in the Unity layer with compilation issues. It seems that, even though none of my code in that class referenced MonoBehavior, the class requires an inheritance from MonoBehavior at runtime. This could possibly be due to using the script as a GameObject elsewhere. Well this put me in a bit of a bind with wanting to also use the PlayerScore from the logic layer. C# of course does not support multiple inheritance.

So I created an interface, IPlayerScore, on the logic layer and had the Logic.PlayerScore implement the interface. From their I made Unity.PlayerScore inherit from MonoBehavior once more, and also implement the interface for Logic.IPlayerScore. Then when the Unity.PlayerScore.Start is triggered it instantiates a new Logic.PlayerScore, and the implementation within Unity.PlayerScore calls the implementations for Logic.PlayerScore. This worked out really well and is super similar to overriding and calling a base class.