Added a generic singleton class for Unity
This ensures that when the object begins its intialization that it will be destroyed if another instance is already present. In addition it will prevent itself from being destroyed if it is the only one in existence. Awake is the first method that is called as part of the execution order, see: https://docs.unity3d.com/Manual/ExecutionOrder.html.