Work in progress, working on resetting the Player when starting a new game
Rather than have the scene loader tell the player to reset directly, I am working on using Unity Events to trigger the action. I have added two new events for when the game starts and whenever a level loads. I have also updated the SceneLoader to be a Singleton now, which was dead simple thanks to my base Singleton classes. I have also updated PlayerHealth and PlayerScore to enable being reset.

Currently I have noticed that the StartMenu no longer has access to the SceneLoader when the scene is recreated. This probably is due to the object being destroyed on awake. I will need to see if there is an easy way to connect the button's onclick event to the SceneLoader during runtime. I am assuming that this means that I will need to make a DataModel for each of the buttons and add listeners similar to what I did with the Player object.

I also imagine that this issue exists in the Game Over screen as well, but we have not gone there twice so I have not been able to see that for myself just yet.