Tinkering with using the Unity Test Framework
I am a strong proponent of writing unit tests to ensure that you do not develop regressions and that the software behaves as expected once exercised. This is why I have written some of the game within a business layer. But as the game has grown in complexity I realize that I will need something to help with the testing within the game layer. Unfortunately the Unity official documentation is rather limited and what I have included so far covers all of what they have covered.