Massive refactor that I did using a second repository.
I had trouble getting the game to continue properly after a game over occurred. To solve this I rebuilt the entire game from scratch and to avoid the previous pitfalls I actually introduced unit testing for most of the Unity scripts. This enabled me to really dial in the functionality and to ensure that the game was stable as I added in additional features.