Adding explosion element
I have added a sprite sheet, sliced it, and converted into an animation. Then I added a data model for the explosion that will destroy it once it finishes playing. Finally I added an explosion audio sound. All of this was converted into a prefab, next steps will be to introduce a variable prefab for the Block. I think I will color it Red, and once it is destroyed it will trigger the explosion.

I think I will need to see if I can add a collider to the explosion so that it can damage the blocks for me, rather than have to damage them directly in code.