Explosion now damages the bricks
This was actually pretty simple to implement. I just had to add a RigidBody2d and a CircleCollider to the explosion prefab. I set the body type of the RigidBody2D to be Dynamic, when set to Kinematic the blocks were not triggering the collision. Unfortunately Dynamic then applies gravity to the object so you could see it falling, this was less than ideal. Luckily Unity had an option to lock the X and Y position of the object, setting this still allows the collisions to be triggered.