Adding a Testing Ground and created a block variant for explosions
Created a testing ground so that I do not have to keep testing on levels that I consider to be production ready. Also added a variant to the original block that is ready and contains a tag called explosion. I will next update the Block so that it has an event that is fired when the block is destroyed. This way I can add an event handler to listen for the block to be destroyed and if the block has a tag for explosion... kaboom.