The red block now explodes.
When the block is destroyed the blockDestroyedEvent is triggered. The BlockExplosionEventHandler, which is only present on Red Blocks, has added a listener for this event. With the event triggered the handler will instantiate a new Explosion. This explosion has a Rigid Body and a Collider so that it can damage the blocks around it. But this also means that the ball could collide with it, which is very silly. So I have implemented some code to ignore collisions between the two objects when the explosion is created.